refactor: rename file and fix typo
