+++ dipu/torch_dipu/csrc_dipu/CMakeLists.txt
@@ -92,7 +92,7 @@ set(TORCH_DIPU_SOURCE
   diopirt/diopirt_impl.cpp
   diopirt/diopi_helper.cpp
 
-  metrics/export.cc
+  metrics/metrics.cpp
 
   profiler/collection.cpp
   profiler/CorrelationIDManager.cpp
