+++ dipu/torch_dipu/csrc_dipu/metrics/metrics.cpp
@@ -1,4 +1,4 @@
-#include "export.h"
+#include "metrics.h"
 
 #include <atomic>
